+++ src/actions/networks.ts
@@ -5,7 +5,7 @@ import {
 import { UttuQuery } from 'api';
 import { deleteNetwork, networkMutation } from 'api/uttu/mutations';
 import { getNetworkByIdQuery, getNetworksQuery } from 'api/uttu/queries';
-import { AppThunk, sentryCaptureException } from 'app/store';
+import { AppThunk, sentryCaptureException } from 'store/store';
 import { UttuError, getStyledUttuError } from 'helpers/uttu';
 import { Network } from 'model/Network';
 import { RECEIVE_NETWORK, RECEIVE_NETWORKS } from './constants';
