+++ src/actions/organisations.ts
@@ -1,6 +1,6 @@
 import { UttuQuery } from 'api';
 import { getOrganisationsQuery } from 'api/uttu/queries';
-import { AppThunk, sentryCaptureException } from 'app/store';
+import { AppThunk, sentryCaptureException } from 'store/store';
 import { OrganisationState } from 'reducers/organisations';
 import { RECEIVE_ORGANISATIONS } from './constants';
