+++ src/config/README.md
@@ -1,0 +1,3 @@
+# Configuration in enki
+
+This document should describe the configuration of enki.
