+++ src/config/configSlice.ts
@@ -1,5 +1,5 @@
 import { createSlice, PayloadAction } from '@reduxjs/toolkit';
-import type { RootState } from 'app/store';
+import type { RootState } from 'store/store';
 import { Config } from 'config/ConfigContext';
 
 export interface ConfigState extends Config {
