+++ src/scenes/App/components/NavBar/UserPreference/index.tsx
@@ -5,7 +5,7 @@ import { sortProviders } from 'model/Provider';
 import { useIntl } from 'react-intl';
 import { useDispatch } from 'react-redux';
 import { useNavigate } from 'react-router-dom';
-import { useAppSelector } from '../../../../../app/hooks';
+import { useAppSelector } from '../../../../../store/hooks';
 import UserMenu from './UserMenu/';
 import './styles.scss';
