+++ src/scenes/FlexibleLines/index.tsx
@@ -9,7 +9,7 @@ import { useEffect, useState } from 'react';
 import { useIntl } from 'react-intl';
 import { useDispatch } from 'react-redux';
 import { Link, useNavigate } from 'react-router-dom';
-import { useAppSelector } from '../../app/hooks';
+import { useAppSelector } from '../../store/hooks';
 import './styles.scss';
 
 export default () => {
