+++ src/store/store.ts
@@ -1,7 +1,7 @@
 import { AnyAction, ThunkAction, configureStore } from '@reduxjs/toolkit';
 import * as Sentry from '@sentry/react';
-import authSlice from 'features/app/authSlice';
-import configSlice from 'features/app/configSlice';
+import authSlice from 'auth/authSlice';
+import configSlice from 'config/configSlice';
 import intlSlice from 'i18n/intlSlice';
 import reducers from 'reducers';
 import immutableStateInvariantMiddleware from 'redux-immutable-state-invariant';
